refactor(agent): merge actions and operations tools into scheduling
Both modules covered halves of the same scheduling domain — jobs, runs,
and backfills — split by read vs write. One scheduling module now groups
the ten tools by domain object, and the matching OperationsAgent +
ActionAgent pair collapses into a single SchedulingAgent carrying both
the monitoring guidance and the confirm-before-acting rules.

By Digitl
